d_drawSize as float

+++ Sofa/Component/Engine/Select/src/sofa/component/engine/select/BaseROI.h
@@ -122,7 +122,7 @@ public:
     Data<bool> d_drawQuads; ///< Draw Quads. (default = false)
     Data<bool> d_drawTetrahedra; ///< Draw Tetrahedra. (default = false)
     Data<bool> d_drawHexahedra; ///< Draw Tetrahedra. (default = false)
-    Data<double> d_drawSize; ///< rendering size for ROI and topological elements
+    Data<float> d_drawSize; ///< rendering size for ROI and topological elements
     Data<bool> d_doUpdate; ///< If true, updates the selection at the beginning of simulation steps. (default = true)
 
     SOFA_ATTRIBUTE_DEPRECATED__RENAME_DATA_IN_ENGINE_SELECT()
